[Performance] Increased performance.
Previously, when opening record occurred, there were multiple attribute sets initialized and merged into the single one.
After profiling it has been revealed that this action was one of the heaviest for record opening.

Now multiple attribute sets is aggregated into single `view` class, and attribute lookup is performed when it is needed by someone.

The code is really dirty at this moment, so it needs some love.