[API] Drop global logger attribute set.
This is breaking change.

After half year of user experience I've realized that nobody uses logging-attached attributes. The main reason of it - because of Wrappers. You can create logger wrapper, attach some attributes and work with it without copying underlying logger resources like file descriptors or sockets.
As a result - you can have only one logger for application, but with multiple wrappers. This is a big advantage.