fix: fixed potential deadlock
There is a RC that appears when the queue is full and its producers
are about to enqueue additional element. Simultaneously the other thread
may empty that queue completely, which results in deadlock when using
blocking wait policy. This commit adds a hack that simply transforms
the CV into timed CV. Otherwise a blocking thead-safe queue is required
or whatever else that I can't imagine right now.